CONTENTdm: Big 2021 update
Addressed PR comments, updated selectors and data source (XML -> JSON).

Most CDM catalogs are idiosyncratic in one way or another -
idiosyncrasies like using the author field to store types of fruit. We
don't just want to give up when we encounter a catalog like that, but we
also don't want to take the fields at face value, or we'll get metadata
that isn't close to useful. To make more educated guesses, we need the
"standardized" field names *and* their labels in the UI. CDM's JSON
gives us both. XML only gives the former. To maintain compatibility with
old versions, we scrape the page instead if JSON is not available.